+++ noodles-sam/src/io/writer/header/record/value/string.rs
@@ -1,8 +1,8 @@
 use std::io::{self, Write};
 
-pub(crate) fn write_string<W>(writer: &mut W, s: &str) -> io::Result<()>
+pub(crate) fn write_string<W>(writer: &mut W, buf: &[u8]) -> io::Result<()>
 where
     W: Write,
 {
-    writer.write_all(s.as_bytes())
+    writer.write_all(buf)
 }